# docs/swagger.yaml
basePath: /
definitions:
...
  history-api_internal_dtos_request.MediaBulkDeleteDto:
    properties:
      media_ids:
        items:
          type: string
        type: array
    required:
    - media_ids
    type: object
  history-api_internal_dtos_request.PreSignedCompleteDto:
    properties:
      token_id:
        type: string
    required:
    - token_id
    type: object
...
  history-api_internal_dtos_response.CommonResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
  history-api_internal_dtos_response.PaginatedResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/history-api_internal_dtos_response.PaginationMeta'
      status:
        type: boolean
    type: object
  history-api_internal_dtos_response.PaginationMeta:
    properties:
      current_page:
        type: integer
      page_size:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for History API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: History API
  version: "1.0"
paths:
...
  /media:
    delete:
      consumes:
      - application/json
      description: Delete multiple media files by IDs
      parameters:
      - description: Media IDs to delete
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/history-api_internal_dtos_request.MediaBulkDeleteDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Delete media
      tags:
      - Media
    get:
      consumes:
      - application/json
      description: Search media with filters, pagination
      parameters:
      - in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - in: query
        minimum: 0
        name: max_size
        type: integer
      - in: query
        maxLength: 100
        name: mime_type
        type: string
      - in: query
        minimum: 0
        name: min_size
        type: integer
      - enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - in: query
        minimum: 1
        name: page
        type: integer
      - in: query
        maxLength: 200
        minLength: 2
        name: search
        type: string
      - enum:
        - id
        - created_at
        - updated_at
        - size
        - original_name
        - storage_key
        - mime_type
        in: query
        name: sort
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: user_ids
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Search media
      tags:
      - Media
  /media/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a media file by ID
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Delete media
      tags:
      - Media
    get:
      consumes:
      - application/json
      description: Retrieve a media file by its ID
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Get media by ID
      tags:
      - Media
  /media/presigned:
    get:
      consumes:
      - application/json
      description: Generate a presigned URL for direct upload to storage
      parameters:
      - in: query
        name: content_type
        required: true
        type: string
      - in: query
        name: fileName
        required: true
        type: string
      - in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Generate presigned URL
      tags:
      - Media
  /media/presigned/complete:
    post:
      consumes:
      - application/json
      description: Confirm that upload via presigned URL is completed
      parameters:
      - description: Request body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/history-api_internal_dtos_request.PreSignedCompleteDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Confirm presigned upload
      tags:
      - Media
  /media/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload media file through server
      parameters:
      - description: Upload file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/history-api_internal_dtos_response.CommonResponse'
      security:
      - BearerAuth: []
      summary: Upload media (server-side)
      tags:
      - Media
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"